fix test

+++ src/software/jetson_nano/services/network/network.h
@@ -42,7 +42,7 @@ class NetworkService
     // Constants
     static constexpr float PROTO_LOSS_WARNING_THRESHOLD          = 0.1f;
     static constexpr unsigned int ROBOT_STATUS_BROADCAST_RATE_HZ = 30;
-    static constexpr unsigned int ROBOT_STATUS_TO_THUNDERLOOP_HZ_RATIO =
+    static constexpr double ROBOT_STATUS_TO_THUNDERLOOP_HZ_RATIO =
         ROBOT_STATUS_BROADCAST_RATE_HZ / (CONTROL_LOOP_HZ + 1.0);
 
     // Variables
